First version of Eudat search; with mapping to schema.org types
